[FIX] public_budget: Fix in budget position analisys view
Only occurs in the budget to date when you visualize and try to analice the budget position with the button.
The values that you see are in zero, this is because the "active_id" that use in the context of button (to set what budget is from) it was not the budget "ID" it was the wizard "ID".