[FIX] public_budget: bug to budget at date (#329)
This solve an error when filter by payment_date and not take account that probably have some payment cancel without payment date.
So we filter also the state and the payment date in the same line. This apply too the confirmation_date in the caculation of the to pay amount values.