[FIX] public_budget: The invoice isn't paid when validate on the wizard
Since for avoid problems with contrains we force the change of account of the line using _write.
The problems is when try to validate the invoice created in the wizard of create invoice, the change of the account is not apply yet intil the transacction it's over. And the transacction is over when the invoice is validated and the state is posted but not paid for this problem.