[13.0] [FIX] public_budget: Validate advance return with the same name of entry
The problem occurs when wanting to use the same name of the advance for the entry name, as the name of the advance is repeated, there is already a generated entry with that name. Odoo restrict duplicated name of entry for same journal and type.
What we do is use the ref as part of the name and we do not complete the ref field since the display name already completes it and we would be repeating the same number.